ci(playwright): build+start in CI, add Edge-only local project; tests hardened
