+++ apps/frontend/package.json
@@ -7,7 +7,8 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "test": "playwright test"
+    "test": "playwright test",
+    "test:edge": "cross-env PW_USE_SYSTEM_EDGE=1 playwright test --project=edge"
   },
   "dependencies": {
     "framer-motion": "^12.23.12",
@@ -17,14 +18,15 @@
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
+    "@playwright/test": "^1.42.0",
     "@tailwindcss/postcss": "^4",
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
+    "cross-env": "^10.0.0",
     "eslint": "^9",
     "eslint-config-next": "15.4.6",
     "tailwindcss": "^4",
-    "typescript": "^5",
-    "@playwright/test": "^1.42.0"
+    "typescript": "^5"
   }
 }
